Añadido modificación usuario y CategoriaControlador

+++ src/main/java/com/proyectointegrado/skillswap/controladores/CategoriaControlador.java
@@ -1,8 +1,13 @@
 package com.proyectointegrado.skillswap.controladores;
 
+import com.proyectointegrado.skillswap.DTOs.CategoriaRequestDTO;
 import com.proyectointegrado.skillswap.entidades.Categoria;
+import com.proyectointegrado.skillswap.entidades.Role;
+import com.proyectointegrado.skillswap.entidades.Usuario;
 import com.proyectointegrado.skillswap.servicios.CategoriaServicio;
+import org.modelmapper.ModelMapper;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -21,17 +26,27 @@ public class CategoriaControlador {
     }
 
     @PostMapping
-    public ResponseEntity<?> crearCategoria(@RequestBody Categoria categoria) {
+    public ResponseEntity<?> crearCategoria(@AuthenticationPrincipal Usuario usuario, @RequestBody CategoriaRequestDTO categoriaRequestDTO) {
+        if (usuario == null || !usuario.getRol().equals(Role.ADMIN)) {
+            return ResponseEntity.status(401).body("No tienes permisos para crear una categoría");
+        }
 
+        Categoria categoria = new Categoria();
+        ModelMapper modelMapper = new ModelMapper();
+        modelMapper.map(categoriaRequestDTO, categoria);
+
+        categoriaServicio.guardarCategoria(categoria);
 //        TODO comprobar que hay usuario logado y que es admin
 
-        return ResponseEntity.ok(categoriaServicio.guardarCategoria(categoria));
+        return ResponseEntity.ok(categoria);
     }
 
 //    TODO comprobar lo del ID si está bien
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> eliminarCategoria(@RequestParam Long id) {
-//        TODO comprobar que hay usuario logado y que es admin
+    public ResponseEntity<?> eliminarCategoria(@AuthenticationPrincipal Usuario usuario, @PathVariable Long id) {
+        if (usuario == null || !usuario.getRol().equals(Role.ADMIN)) {
+            return ResponseEntity.status(401).body("No tienes permisos para crear una categoría");
+        }
 
         categoriaServicio.eliminarCategoria(id);
